requirements.txt
crewai
openai
streamlit
requests
hypothesis
python-dotenv
